Fixed error where no private files are deleted on import

+++ lib/src/data/repositories/io_repository.dart
@@ -21,11 +21,13 @@ class IoRepository implements IoRepositoryInterface {
   }
 
   @override
-  Future<String> readFromTextFile(String pathToFile) async {
+  Future<String> readFromTextFile(String pathToFile, bool privateFile) async {
     String result = '';
     File file = File(pathToFile);
     result = await file.readAsString();
-    file.delete();
+    if (privateFile) {
+      file.delete();
+    }
     return result;
   }
